Let stow deal with all symlinks
The install_pre_commit_hook function would only pass if the symlink does
not exist (i.e. the first time it is run). Stow handles symlinks better,
lets use it. Furthermore adding the warn-detached-head file to the `git`
bundle means it is going to appear as a symlink in the homedir which is
not desirable