Add a sane mapping for COQ's jump_to_mark
This is used to jump between placeholders (i.e. function arguments)
after a completion selection is made. I use this only in insert
mode, and I'd like a nice binding that makes sense. COQ however will
always remap this in normal mode as well, which is a bit annoying. It
also seems to do this when it is loading. To combat this the only option
I've worked out so far is to do a autocmd that remaps this key to the
vin-tmux-navigator binding on each BufEnter