Improve gopls memory consumption and use hard limit only
Using the high memory cgroup v2 option causes gopls to be too
unresponsive to use. Additionally 1/3 of my total memory (16G) isn't
enough for it. Using the `DegradeClosed` memory mode improves the
consuption a bit and also using a hard limit of 1/2 of my memory seems
to perform better for now.